Crate wgpu_test

This module contains common test-only code that needs to be shared between the examples and the tests.

Re-exports

pub use self::image::ComparisonType;

Modules

image

Structs

FailureCase

Conditions under which a test should fail or be skipped.

FailureReasons

SurfaceGuard

TestParameters

TestingContext

Functions

fail

fail_if

initialize_instance

initialize_test

valid