wgpu_test

Module expectations

Structs

FailureApplicationReasons

Reason why a test matches a given failure case.

FailureCase

Conditions under which a test should fail or be skipped.

FailureReason

Reason why a test is expected to fail.

FailureResult 🔒

Enums

ExpectationMatchResult 🔒

FailureBehavior

FailureResultKind 🔒

Functions

expectations_match_failures 🔒

Compares if the actual failures match the expected failures.