wgpu_test
Module init

Structs

SurfaceGuard

Functions

init_logger

Initialize the logger for the test runner.

initialize_adapter

Initialize a wgpu adapter, using the given adapter report to match the adapter.

initialize_device

Initialize a wgpu device from a given adapter.

initialize_instance

Initialize a wgpu instance with the options from the environment.