wgpu_test::init

Function default_device_lost_callback

fn default_device_lost_callback(reason: DeviceLostReason, message: String)

Default device-lost callback installed by initialize_device. Panics on any non-wgpu::DeviceLostReason::Destroyed device loss, which will cause the test to be treated as a failure.

Tests intentionally provoking device loss should install their own callback with wgpu::Device::set_device_lost_callback.