wgpu_test

Module isolation

Structs

OneTestPerProcessGuard

When this guard is active, enforces that there is only a single test running in the process at any one time. If there are multiple processes, creating the guard hard terminates the process.

Constants

OTHER_TEST_IN_PROGRESS_ERROR 🔒

Statics

TEST_ACTIVE_IN_PROCESS 🔒

True if a test is in progress somewhere in the process, false otherwise.