wgpu_test

Struct TestParameters

pub struct TestParameters {
    pub required_features: Features,
    pub required_downlevel_caps: DownlevelCapabilities,
    pub required_limits: Limits,
    pub required_instance_flags: InstanceFlags,
    pub force_fxc: bool,
    pub skips: Vec<FailureCase>,
    pub failures: Vec<FailureCase>,
}

This information determines if a test should run.

Fields
required_features: Featuresrequired_downlevel_caps: DownlevelCapabilitiesrequired_limits: Limitsrequired_instance_flags: InstanceFlagsforce_fxc: bool

On Dx12, specifically test against the Fxc compiler.

For testing workarounds to Fxc bugs.

skips: Vec<FailureCase>

Conditions under which this test should be skipped.

failures: Vec<FailureCase>

Conditions under which this test should be run, but is expected to fail.

Implementations

impl TestParameters

pub fn test_features_limits(self) -> Self

Set of common features that most internal tests require for compute and readback.

pub fn features(self, features: Features) -> Self

Set the list of features this test requires.

pub fn downlevel_flags(self, downlevel_flags: DownlevelFlags) -> Self

pub fn limits(self, limits: Limits) -> Self

Set the limits needed for the test.

pub fn instance_flags(self, instance_flags: InstanceFlags) -> Self

Sets the instance flags that the test requires.

pub fn force_fxc(self, force_fxc: bool) -> Self

pub fn expect_fail(self, when: FailureCase) -> Self

Mark the test as always failing, but not to be skipped.

pub fn skip(self, when: FailureCase) -> Self

Mark the test as always failing, and needing to be skipped.

pub fn enable_noop(self) -> Self

Enable testing against the noop backend and miri.

The noop backend does not execute any operations, but allows us to test validation and memory safety.