Crate wgpu_test

Test utilities for the wgpu repository.

Re-exports

pub use self::image::ComparisonType;

Modules

image

Image comparison utilities

native

Infrastructure for the native, cargo-nextest based harness.

Macros

gpu_test_main

Adds the necissary main function for our gpu test harness.

Structs

FailureApplicationReasons

Reason why a test matches a given failure case.

FailureCase

Conditions under which a test should fail or be skipped.

FailureReason

Reason why a test is expected to fail.

GpuTestConfiguration

Configuration for a GPU test.

TestParameters

This information determines if a test should run.

TestingContext

Parameters and resources hadned to the test function.

Enums

FailureBehavior

Functions

did_fail

Returns true if the provided callback fails validation.

execute_test

Execute the given test configuration with the given adapter report.

fail

Run some code in an error scope and assert that validation fails.

fail_if

Run some code in an error scope and assert that validation succeeds or fails depending on the provided should_fail boolean.

initialize_adapter

Initialize a wgpu adapter, using the given adapter report to match the adapter.

initialize_device

Initialize a wgpu device from a given adapter.

initialize_instance

Initialize a wgpu instance with the options from the environment.

valid

Run some code in an error scope and assert that validation succeeds.

Attribute Macros

gpu_test

Creates a test that will run on all gpus on a given system.