Module wgpu_test::image

Image comparison utilities

Structs

ReadbackBuffers

Enums

ComparisonType

The FLIP library generates a per-pixel error map where 0.0 represents “no error” and 1.0 represents “maximum error” between the images. This is then put into a weighted-histogram, which we query to determine if the errors between the test and reference image is high enough to count as “different”.

Functions

compare_image_output